Remove wincompatliband reorganize compat section of the lib
